daemon/requirements.txt
unum-ledger @ git+https://github.com/unum-apps/ledger@0.1.3
discord.py==2.4.0
atproto==0.0.56
micro-logger==0.1.2
prometheus-client==0.12.0
redis==5.2.1
ptvsd==4.3.2
coverage==5.2.1
pylint==2.5.3
